ui: Add shuffle functionality
Use a queue object to hold the songs and shuffle that when required. The
implementation is simplistic and currently assumes that the queue always
contains the full library of songs. This can be improved to hold only song
names, when we add the ability to add/remove items from the queue.