ui: Don't change window.location.hash when playing songs
We now have a way to copy the link to a specific song. Changing the hash with
the currently playing song doesn't seem necessary now. It may be nice to save
the state of the page to allow for reloads, etc. but the hash can get quite
long and ugly especially when using a URL for the song src. We can add back
some kind of state to the URL on the current song, when the hash format is a
little more steamlined.